v5: add tests
