+++ tests/numbers.test.ts
@@ -55,4 +55,155 @@ describe('Numbers', () => {
         .with(P.number, () => 'number')
         .exhaustive();
   });
+
+  describe('chainable', () => {
+    it(`P.number.between(1, 10)`, () => {
+      const f = (input: string | number) =>
+        match(input)
+          .with(P.number.between(0, 10), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'between 0 and 10';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number>>;
+            return 'something else';
+          });
+
+      expect(f(5)).toBe('between 0 and 10');
+      expect(f(0)).toBe('between 0 and 10');
+      expect(f(10)).toBe('between 0 and 10');
+      expect(f('gabriel')).toBe('something else');
+    });
+
+    it(`P.number.lt(..)`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.lt(10), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('yes');
+      expect(f(12)).toBe('no');
+      expect(f(10n)).toBe('no');
+    });
+    it(`P.number.gt(..)`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.gt(10), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('no');
+      expect(f(10)).toBe('no');
+      expect(f(12)).toBe('yes');
+    });
+    it(`P.number.gte(..)`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.gte(10), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('no');
+      expect(f(10)).toBe('yes');
+      expect(f(12)).toBe('yes');
+    });
+    it(`P.number.lte(..)`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.lte(10), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('yes');
+      expect(f(10)).toBe('yes');
+      expect(f(12)).toBe('no');
+    });
+    it(`P.number.int(..)`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.int(), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('yes');
+      expect(f(10.123)).toBe('no');
+      expect(f(-Infinity)).toBe('no');
+    });
+    it(`P.number.finite()`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.finite(), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('yes');
+      expect(f(10.123)).toBe('yes');
+      expect(f(-Infinity)).toBe('no');
+    });
+    it(`P.number.positive()`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.positive(), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('yes');
+      expect(f(10.123)).toBe('yes');
+      expect(f(-10.123)).toBe('no');
+      expect(f(-Infinity)).toBe('no');
+    });
+    it(`P.number.negative()`, () => {
+      const f = (input: string | number | bigint) =>
+        match(input)
+          .with(P.number.negative(), (value) => {
+            type t = Expect<Equal<typeof value, number>>;
+            return 'yes';
+          })
+          .otherwise((value) => {
+            type t = Expect<Equal<typeof value, string | number | bigint>>;
+            return 'no';
+          });
+
+      expect(f(5)).toBe('no');
+      expect(f(10.123)).toBe('no');
+      expect(f(-10.123)).toBe('yes');
+      expect(f(-Infinity)).toBe('yes');
+    });
+  });
 });
